feat(ledger): alter step columns to BIGINT (m20260510, closes Wave 23 drift) (#119)
* feat(migration): add m20260510_000001_step_to_bigint (Wave 24)

Adds idempotent migration to ALTER step columns to BIGINT:
  - public.bpb_samples.step       INT -> BIGINT
  - ssot.bpb_samples.step         INT -> BIGINT (guarded IF EXISTS)
  - public.igla_race_trials.final_step  INT -> BIGINT
  - public.igla_race_trials.steps_done  BIGINT already (no-op guard)

All ALTER statements use DO $$ ... END $$; guards via information_schema
so re-running on an already-migrated DB is a safe no-op.

down() is an intentional no-op: reversing BIGINT to INT is lossy
when values > INT_MAX exist.

Registers migration in Migrator::migrations() after init_schema.
Adds Wave 24 idempotency test (ignored without DATABASE_URL).

Anchor: phi^2 + phi^-2 = 3 · DOI 10.5281/zenodo.19227877
Closes #118

* feat(entities): lift step/final_step to i64 (Wave 24 BIGINT migration)

- bpb_samples::Model.step: i32 -> i64 (matches new BIGINT column)
- igla_race_trials::Model.final_step: Option<i32> -> Option<i64>
- steps_done already i64, no change needed
- Updated doc-comment in bpb_samples.rs: step INT -> step BIGINT

Anchor: phi^2 + phi^-2 = 3 · DOI 10.5281/zenodo.19227877
Refs #118

* fix(writer): cast step to i64 in bpb_sample ActiveModel (Wave 24)

- neon_writer::bpb_sample: step: Set(step) -> Set(step as i64)
  to match the new BIGINT column type
- tests/ledger_seaorm.rs: step: i32 = 200 -> step: i64 = 200
  to match updated Model type

Public function signature bpb_sample(seed: i32, step: i32, ...) is
unchanged for callers; the cast happens internally.

Anchor: phi^2 + phi^-2 = 3 · DOI 10.5281/zenodo.19227877
Refs #118

* docs(changelog): append Wave 24 STEP-TO-BIGINT entry to MIGRATION.md

Anchor: phi^2 + phi^-2 = 3 · DOI 10.5281/zenodo.19227877
Refs #118

---------

+++ src/entities/bpb_samples.rs
@@ -3,7 +3,7 @@
 // SeaORM entity for public.bpb_samples (legacy ledger table).
 //
 // Column types match the DDL in /tmp/full_ddl.sql:
-//   id BIGSERIAL, canon_name TEXT, seed BIGINT, step INT, bpb DOUBLE PRECISION,
+//   id BIGSERIAL, canon_name TEXT, seed BIGINT, step BIGINT, bpb DOUBLE PRECISION,
 //   ts TIMESTAMPTZ
 //
 // Note: seed is BIGINT -> i64 in Rust (fixes the i32/INT8 mismatch from #114).
@@ -18,7 +18,7 @@ pub struct Model {
     pub canon_name: String,
     /// BIGINT — must be i64, not i32 (fixes #114).
     pub seed: i64,
-    pub step: i32,
+    pub step: i64,
     pub bpb: f64,
     pub ts: DateTimeWithTimeZone,
 }
